+++ lib/whois/record/parser/whois.pnina.ps.rb
@@ -7,73 +7,25 @@
 #++
 
 
-require 'whois/record/parser/base'
+require 'whois/record/parser/base_cocca'
 
 
 module Whois
   class Record
     class Parser
 
-      #
-      # = whois.pnina.ps parser
-      #
       # Parser for the whois.pnina.ps server.
       #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
       #
-      class WhoisPninaPs < Base
-
-        property_supported :status do
-          if content_for_scanner =~ /Status:\s+(.+?)\n/
-            case (s = $1.downcase)
-            when /^active/
-              :registered
-            when "not registered"
-              :available
-            else
-              Whois.bug!(ParserError, "Unknown status `#{s}'.")
-            end
-          else
-            Whois.bug!(ParserError, "Unable to parse status.")
-          end
-        end
-
-        property_supported :available? do
-          status == :available
-        end
-
-        property_supported :registered? do
-          !available?
-        end
-
-
-        property_supported :created_on do
-          if content_for_scanner =~ /Created:\s+(.+?)\n/
-            Time.parse($1)
-          end
-        end
-
-        property_not_supported :updated_on
-
-        property_supported :expires_on do
-          if content_for_scanner =~ /Expires:\s+(.+?)\n/
-            Time.parse($1)
-          end
-        end
-
-
-        property_supported :nameservers do
-          if content_for_scanner =~ /Name Servers:\n((.+\n)+)\n/
-            $1.split("\n").delete_if { |value| value =~ /nodns\.ns$/i  }.map do |name|
-              Record::Nameserver.new(:name => name.strip)
-            end
-          end
-        end
-
+      class WhoisPninaPs < BaseCocca
+        self.status_mapping = {
+          "active" => :registered,
+          "delegated" => :registered,
+          "not registered" => :available,
+          "active - pending transfer" => :registered,
+        }
       end
 
     end