+++ spec/whois/record/parser/responses/whois.centralnic.com/jpn.com/status_registered_spec.rb
@@ -38,7 +38,7 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   end
   describe "#status" do
     it do
-      subject.status.should == ["OK"]
+      subject.status.should == ["OK", "RENEW PERIOD"]
     end
   end
   describe "#available?" do
@@ -66,7 +66,7 @@ describe Whois::Record::Parser::WhoisCentralnicCom, "status_registered.expected"
   describe "#expires_on" do
     it do
       subject.expires_on.should be_a(Time)
-      subject.expires_on.should == Time.parse("2013-06-29 23:59:59 UTC")
+      subject.expires_on.should == Time.parse("2015-06-29 23:59:59 UTC")
     end
   end
   describe "#registrar" do