+++ spec/whois/record/parser/responses/whois.gg/status_registered_spec.rb
@@ -21,6 +21,16 @@ describe Whois::Record::Parser::WhoisGg, "status_registered.expected" do
     described_class.new(part)
   end
 
+  describe "#domain" do
+    it do
+      subject.domain.should == "google.gg"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      lambda { subject.domain_id }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
   describe "#status" do
     it do
       subject.status.should == :registered
@@ -39,18 +49,17 @@ describe Whois::Record::Parser::WhoisGg, "status_registered.expected" do
   describe "#created_on" do
     it do
       subject.created_on.should be_a(Time)
-      subject.created_on.should == Time.parse("2003-04-30 00:00:00")
+      subject.created_on.should == Time.parse("2003-04-30 00:00:00 UTC")
     end
   end
   describe "#updated_on" do
     it do
-      subject.updated_on.should be_a(Time)
-      subject.updated_on.should == Time.parse("2011-03-31 00:00:00")
+      subject.updated_on.should == nil
     end
   end
   describe "#expires_on" do
     it do
-      lambda { subject.expires_on }.should raise_error(Whois::AttributeNotSupported)
+      subject.expires_on.should == nil
     end
   end
   describe "#nameservers" do