+++ spec/whois/record/parser/responses/whois.nic.as/status_available_spec.rb
@@ -21,6 +21,16 @@ describe Whois::Record::Parser::WhoisNicAs, "status_available.expected" do
     described_class.new(part)
   end
 
+  describe "#domain" do
+    it do
+      subject.domain.should == "u34jedzcq.as"
+    end
+  end
+  describe "#domain_id" do
+    it do
+      lambda { subject.domain_id }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
   describe "#status" do
     it do
       subject.status.should == :available
@@ -38,17 +48,22 @@ describe Whois::Record::Parser::WhoisNicAs, "status_available.expected" do
   end
   describe "#created_on" do
     it do
-      lambda { subject.created_on }.should raise_error(Whois::AttributeNotSupported)
+      subject.created_on.should == nil
     end
   end
   describe "#updated_on" do
     it do
-      lambda { subject.updated_on }.should raise_error(Whois::AttributeNotSupported)
+      subject.updated_on.should == nil
     end
   end
   describe "#expires_on" do
     it do
-      lambda { subject.expires_on }.should raise_error(Whois::AttributeNotSupported)
+      subject.expires_on.should == nil
+    end
+  end
+  describe "#registrar" do
+    it do
+      subject.registrar.should == nil
     end
   end
   describe "#nameservers" do