+++ spec/whois/record/parser/responses/whois.nic.bo/status_registered_spec.rb
@@ -60,7 +60,7 @@ describe Whois::Record::Parser::WhoisNicBo, "status_registered.expected" do
   describe "#expires_on" do
     it do
       subject.expires_on.should be_a(Time)
-      subject.expires_on.should == Time.parse("2011-08-22")
+      subject.expires_on.should == Time.parse("2013-08-22")
     end
   end
   describe "#nameservers" do