+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 - CHANGED: Updated whois.nic.hn parsers to the new response format.
 
+- CHANGED: Updated whois.nic.as parsers to the new response format.
+
+- CHANGED: Updated whois.nic.mg parsers to the new response format.
+
+- CHANGED: Updated whois.nic.cd parsers to the new response format.
+
+- CHANGED: Updated whois.je parsers to the new response format.
+
+- CHANGED: Updated whois.gg parsers to the new response format.
+
+- CHANGED: Updated whois.pnina.ps parsers to the new response format.
 
 
 ## Release 3.2.0